refactor: (상품) 응답 DTO 추가

+++ src/main/java/kr/hhplus/be/server/product/application/dto/ProductResponseDto.java
@@ -1,12 +1,14 @@
 package kr.hhplus.be.server.product.application.dto;
 
+import kr.hhplus.be.server.order.application.dto.OrderItemRequestDto;
+import kr.hhplus.be.server.order.domain.OrderItem;
 import kr.hhplus.be.server.product.domain.Product;
 import java.math.BigDecimal;
 
 public record ProductResponseDto(
     Long id,
     BigDecimal price,
-    Long stock
+    Integer stock
 ) {
     public static ProductResponseDto from(Product product) {
         return new ProductResponseDto(
